(MODULES-10454) Install latest only when no agent is present
This modifies the behavior of the `puppet_agent::install` task so that
it only assumes that the latest version of the agent should be installed
when no agent is present on the system. Previously, if the `version`
parameter was undefined, the task would default to the latest version
and install it on the system, even if an agent was already present. This
would clobber any existing version of the agent, even unintentionally.