(MODULES-10514) Remove use of `version_powershell` task in `install_powershell`
This removes the use of the `version_powershell` task in the
`install_powershell` task. Previously, this task was used to determine
whether the `puppet-agent` package was already installed. However, some
features used to run the script and parse its output are not available
in versions of powershell older than 3.0.

The `install_powershell` task instead uses behavior similar to the
`version_powershell` task by checking the system's registry entries to
check if the `puppet-agent` package is installed.