(MODULES-10589) Exit early when puppet.list config file has been modified
This updates the `install_shell` task to check whether the configuration
file `/etc/apt/repos.list.d/puppet.list' has been modified prior to
installing the `puppet-release` repo.

Previously, if a user had modified the file (such as by manually
maintaining the repo), `dpkg` would prompt the user whether they wanted
to keep the old config file or replace it with the maintainer's config
file. This caused the script to stall as it waited for user input. While
`dpkg` can answer these prompts non-interactively using the
`--force-conf*` options, this behavior would be undesirable as it could
result in the script making a critical decision for the user. Instead,
the script will provide a helpful warning that the config file has been
modified and that agent installation will be skipped.