(maint) Ensure apt-transport-https is installed before installing agent
The `puppet5` apt repo was recently updated to pull from the https site.
In order to install the `puppet-agent` package from this collection, the
`apt-transport-https` package must be installed on the target. This
updates the `puppet_agent::install` task to check if this package is
installed and install it if it is not.