(MODULES-10653) Failed to upgrade agent using puppet task
Before this commit, when trying to run the
`puppet_agent::install_powershell` task, with a specific version as
parameter, it was failing with below error message:
Error: Timed out waiting for status response from <node_name>

On the node, both `puppet agent` and `pxp-agent` services got stopped
and the event log viewer showed:
Application or service 'task_wrapper' could not be shut down.

Due to Windows agents becoming unresponsive when running this task and
as per MODULES-10633 discussions, the `puppet_agent::install_poweshell`
task should not be used for upgrading Puppet Agents while either of
`puppet agent` or `pxp-agent` services are still running. This Puppet
Agent module task will now fail immediately and output a message if any
of the two affected services are still running.

The task will stop and output a message if the given version is already
installed on the node, thus avoiding unnecessary msi
download/installation. This behaviour is aligned with the existing
implementation for no version given and Puppet Agent already installed
on the node.