(MODULES-10666) Stop agent run after an upgrade
Previously, when evaluating additional resources after a puppet-agent
upgrade, we could get to a point where we had a mixture of old/new
Ruby code loaded in memory. This may cause unexpected behavior if the
new code tries to execute something not available in the already
loaded old code.

This commit introduces a new type/provider that is evaluated whenever
a puppet-agent upgrade happens on a platform managed with the package
resource (AIX, Debian, EL, Suse).

The newly added resource attempts to figure out if a Puppet upgrade
was performed, then stops the run altogether. If Puppet is being run
as a service, the HUP signal will have the daemon trigger a subsequent
Puppet run that will evaluate the remaining resources.

When running `puppet agent -t` manually, it is up to the user to rerun
the command after the puppet-agent upgrade.