(MODULES-10768) Add task and plan for running Puppet agent
This adds a new `puppet_agent::run` task that runs the Puppet agent on a
list of targets, and returns the run report contents as well as the
exit code for the agent run.

This also adds a `puppet_agent::run` plan which checks if the agent is
installed on each target and runs the `puppet_agent::run` task on each
target that has the agent installed. The plan returns a `ResultSet` that
includes fail results for targets that do not have the agent installed,
as well as the results from the `puppet_agent::run` task.