(MODULES-10813) Mismatched versions stops install
When performing a puppet-agent upgrade with PE, the .msi package
that will get downloaded by default is the puppet-agent version
of the server node.

If a different version than the one on the master is requested,
and if no pe_repo is configured and available, it will install the version
on the master.

This PR verifies if the desired version and version from the .msi match.
If they do not match the install_puppet.ps1 will fail before installing
an unwanted version.

This PR also enables downgrading puppet-agent versions on windows if
pe_repo and package version from puppet-agent module match.

More details in https://tickets.puppetlabs.com/browse/MODULES-10813