(MODULES-10822) Fix teardown on Windows
Uninstalling puppet-agent after it has been upgraded can leave orphaned
files on the system.

We encountered this when upgrading from 5.5.16 to 5.5.21, the difference
being that the latter no longer ships with curl.exe. This made the
upgrade keep curl.exe from the 5.5.16 installation, and when
uninstalling 5.5.21, the executable and other libraries would still be
there.

To fix this, remove the entire installation directory after uninstalling
puppet-agent.