(MODULES-10851) Fix Windows nightly prerequisites check
MODULES-10813 introduced a new powershell script that is executed on
every Windows upgrade: `prerequisites_check.ps1`.

This script checks whether the version in the MSI matches what
`$package_version` we want to install.

To install a nightly build, the user would pass something like
`7.0.0.463.gf7e6640e` to `package_version`. This gets munged by the
puppet_agent module into `7.0.0.463` which is then given to the
`prerequisites_check.ps1` script.

It appears that MSI packaging does not discern between released builds
and nightly builds, as the versions reported by the MSI package are
always of type `MAJOR.MINOR.PATCH`, so the script would compare `7.0.0`
with `7.0.0.463`.

Change the script to always strip the commit number from the version.