(MODULES-10879) Implement configuration management
This commit introduces a parameter that can be used to manage agent
configuration. This parameter is restricted to managing only
specifically permitted settings.

Without this functionality, users can only manage agent configuration by
independently writing their own ini_setting resources in their own code.
This is inefficient and error-prone compared to supporting basic
configuration management of the agent in the official Puppet agent
supported module.

By supporting basic configuration management of settings in this module
it will be easier and safer to configure the Puppet agent, as well as
becoming easier to document a correct and maintainable way of doing so.

Tests:

The new puppet_agent::configure class should be included in the catalog
if puppet_agent is managing anything. Tests for puppet_agent validate
this. Tests for puppet_agent::configure functionality are implelmented
in class-specific specs.