(MODULES-10996) Fix SLES 11 PE upgrades
SLES 11 can no longer be upgraded in PE by installing from the repos. To
work around this, if we're SLES 11 and PE, download the package and
install it directly using rpm, regardless of the value of `manage_repo`.
This is the same approach we take for AIX, macOS and Windows. Because
zypper is left in a semi-broken state if our pe_repo is installed, make
sure we remove it and don't install it again.

The GPG keys will continue to be imported.

For non-PE agents, installing from the FOSS repos should still be
possible, so nothing should change on that part. Additional tests were
added to assert the behavior of FOSS vs PE.