(MODULES-10996) Add separate class for suse install
The install class was getting a bit complex, so I pulled out the SUSE
logic into a separate file.

This also made it easier to add a GPG check before installing the RPM
file. Since we don't use zypper anymore on SLES 11 in PE, create a
separate `exec` resource to handle the GPG check before upgrading.