(maint) Update acceptance tests to run without service
By relying on the puppet service to upgrade puppet we lose all logging
info in case something fails, which makes things difficult to debug.
Change to run with puppet agent -t.

We can also fix the logging part by configuring puppet to log to a file,
then printing the contents of the file.