(maint) Readd puppet5 support
We still want puppet5 users who need to upgrade to puppet6 to be able to
use the latest puppet_agent module. Even though the required puppet
version in metadata.json is not enforced when installing the module, it
still shows up on the Forge and can cause confusions as to whether the
module can be used on puppet5 or not.

The bug fixes and improvements that we do in the module still benefit
puppet5 and we should encourage users to always use the latest version
of the module.

This commit updates metadata.json to show puppet5 support.