(MODULES-11066) Add support for running puppet_agent::install in noop mode
This enables running the `puppet_agent::install` task in noop mode. When
the task is run in noop mode, it executes up until the puppet-agent
package is downloaded and installed.

This work enables adding support for running plans in noop mode in Bolt,
which uses the `puppet_agent::install` task as the default
puppet_library plugin when running an `apply_prep`.