(MODULES-11123) Avoid loading puppet facts in `install/windows.pp`
The install puppet command in `manifests/install/windows.pp` is executed
if the desired version is not already installed. This is done by asking
facter the value of the `aio_agent_version` fact on the target node and
is compared to expected package version.

Before this commit, facter was asked to also load puppet facts (using
the `-p` parameter) which was causing high load times due to unnecessary
facts getting loaded and resolved when currently installed puppet agent
version was the only information needed.

This fix removes the puppet facts loading for this command requirement
and lowers catalog application time by as much as half in some
scenarios.