(maint) Allow detection of non-AIO Puppet
When a non-AIO Puppet version is installed (e.g. on FreeBSD where Puppet
packages are maintained by the FreeBSD community), Bolt is unable to
detect the version of Puppet.

If no evidence of the presence of Puppet is found at the AIO locations,
check if a `puppet` command is found in the $PATH, and if so, run it
with the `--version` parameter to get the version of Puppet.