(MODULES-11276) Replace lsbdistcodename with os.disto.codename
Prior to this commit, the debian package name used the lsbdistcodename
legacy fact to determine the package name. That legacy fact requires the
lsb-release package to determine, however in Puppet 7+, the
os.distro.codename fact can be determined without the lsb-release
package installed. This commit replaces the lsbdistcodename with
os.distro.codename for the Debian package name.