(MODULES-11262) Update the version check for acceptance check
When we test upgrades we should only check to look for the major version.
We use to test the latest known puppet-agent package was upgraded. Due
to a bug in our puppet-agent packaging there are times this test would
produce a failure, even though the agent was upgraded to the major version.