(MODULES-11334) Fix TLS error for agent downloads
Most modern sites either require or are starting to require TLS 1.2.
On older Windows systems the default security protocols for PowerShell 5 are SSL3 and TLS
1.0.

If using System.Net.WebClient, Invoke-WebRequest or Invoke-RestMethod to
communicate with a site that has a higher TLS requirement requests will fail with the
following error message:

`Could not create SSL/TLS secure channel.`

This commit adds a new function called Set-Tls12 that will add TLS
1.2 to the list of active security protocols for the current session.

The above will only happen if the Tls12 enum is not present in the
current list.