(maint) replace legacy validate function with datatype
This replaces the validate_absolute_path() function call with a datatype
from puppetlabs/stdlib. The type is available since the 4.13.1 release
of stdlib (and this module already depends on 5.1 and newer).

Some background: I'm currently debugging a slow pe-puppetserver metrics
API. The puppet-profiler is enabled and provides metrics for all called
functions. In our environment the puppet_agent module is the only one
still using this legacy function. By replacing it with a datatype, it
follows current best practice guides and also speeds up the API because
it has to care about less metrics (in my specific case at least).