(MODULES-11361) Updates legacy facts
In preparation for the removal of legacy facts in Puppet 8, this
commit updates all of the legacy facts in code and tests to modern,
structured facts.

This commit also adds the override_facts method to spec_helper_local
in order to support merging in fact overrides into nested hashes of
facts.