(MODULES-11365) Updates Windows nightly tests
Newer versions of Ruby used by the setup-ruby action use the ucrt
C library. Puppet nightly gems are instead built with the mscvcrt
library (AKA mingw), which causes

This commit updates the GitHub Action that tests nightly Puppet gems
to use the generic gem instead of the mingw gem, and adds a step to
install the ffi dependency that was included in the mingw gem but not
the generic gem.