(maint) Unpin Octokit gem
In 99d7dcb we pinned Octokit to 4.21.0 due to an issue with
authentication and the Faraday gem in Octokit 4.22.0. There was
a fix for this issue in 4.23.0, so we no longer need to pin back.

This gem is used to generate changelogs in release prep, and other
modules (i.e. Puppet core modules) have used newer versions of
Octokit (4.25.0) without issue.

This commit removes the pin for Octokit to 4.21.0.