(maint) Delete duplicate tests
This code was last updated in Oct 2020 and all this time we've been testing
"upgrades" from 7.x to 5.5.21. This only recently started to fail when stdlib v9
was released as the root_home fact uses Ruby syntax that's not valid in Puppet
5/Ruby 2.1.9.

It's not practical to update the upgrade tests after each release. And we already test
upgrades and idempotency in the acceptance/tests/*, so delete these tests.