(maint) Block lint checks
This commit blocks the top_scope_facts, only_variable_string, and
strinct_indent lint check for portions of manifests/install/windows.pp,
manifests/osfamily/redhat.pp, and manifests/osfamily/suse.pp respectively.

For the top_scope_facts, the check keeps on thinking
`puppet_agent::_expected_package_version` in windows.pp is a fact. For the
only_variable_string, when I try to remove the double quotes from
`"${facts['os']['release']['major']}"`, the conditional no longer works as
expected in redhat.pp. Lastly, for the strict_indent check, it thinks the
indentation for $script in redhat.pp and suse.pp is wrong but it is not.