(PA-5266) Change checksum algorithm to default
Prior to this commit, the puppet_agent::prepare::package class used the
sha256lite algorithm in the file resource for local installation
packages.

In certain circumstances (e.g. when a download is interrupted), this
checksum, which only checks the first N bytes, does not provide
sufficient information to determine if the file is actually there.

This commit removes the checksum attribute altogether, which causes the
file resource to use its default algorithm, sha256. This algorithm
should handle circumstances like those outlined above.