(PA-5826) Only read Windows VERSION file during puppet apply
Previously, puppetserver attempted to load the wrong VERSION file when managing
`package_version => auto` and compiling catalogs for Windows nodes as can be
seen by the new test:

    Evaluation Error: Error while evaluating a Function Call, Could not find any files from C:\Program Files\Puppet Labs\Puppet\VERSION (file: path/to/init.pp, line: 170, column: 42)

This adds a `puppet_runmode` fact. When running as puppet agent, the run mode will
be `agent`. When running as puppet apply, it will be `user`. If it's the latter
and we're a Windows host, then look for the VERSION file locally.

Since the default class parameter is no longer dependent on the agent
fact, I regenerated the reference using rake strings:generate:reference