(maint) Limit beaker-puppet to Ruby < 3.1
beaker-puppet is not compatible with Ruby >= 3.2 due to taint functions.
If you try to run `bundle install` with Ruby > 3.1 in its gemspec and
beaker-puppet, Bundler errors with:

Because beaker-puppet >= 3.0.0 depends on Ruby >= 2.7, < 3.2
  and Gemfile depends on beaker-puppet ~> 3.0,
  Ruby >= 2.7, < 3.2 is required.
So, because current Ruby version is = 3.2.3,
  version solving has failed.

This commit updates the Gemfile to only list beaker-puppet if Ruby is
less than 3.1.0.