Install puppetcore* rpms from yum-puppetcore.puppet.com
Adds support for puppetcore* collections to the install task.

Adds optional username and password parameters. The username defaults to
'forge-key' and the password must be specified as a forge API key.

When installing the puppetcore* collection, the task will download a release
package from yum-puppetcore.puppet.com/public and add the credentials to the
repo. For most RPM platforms, this is done by adding credentials to the baseurl.

For SLES, the credentials are added to /etc/zypp/credentials.d/PuppetcoreCreds.

Create dnf and sles Dockerfiles for testing the install task.

Create install.sh script to build docker image and run it:

    docker/bin/install.sh [image] [version]

By default install 8.11.0 on rocky8.

The `PUPPET_FORGE_TOKEN` environment variable must be set, which will be passed
as the `password` to the task.