Checkout the latest tagged module version
Checkout the latest tagged module version when installing the agent.
This has to be done separately for dnf & sles. We can't use `puppet
module install` in the install task, because puppet hasn't been
installed yet.

As a result of this change, we have to clone the entire repo, rather
than using depth=1, since there's no guarantee that the latest commit
will be tagged.

Also update the commented out git clone commands in the upgrade scripts.
Those can be useful if you need to test that an unreleased change in a
module dependency won't break this module.