Install puppetcore* dpkg from apt-puppetcore.puppet.com
Adds support for puppetcore* collections to the apt-based install task.

When installing the puppetcore* collection, the task will download a release
package from apt-puppetcore.puppet.com/public and add the credentials to
/etc/apt/auth.conf.d/apt-puppetcore-puppet.conf. The other collections are
unaffected.

The `PUPPET_FORGE_TOKEN` environment variable must be set, which will be passed
as the `password` to the task.

Add debian 10-12 and ubuntu 18.04-24.04 to test installs in docker. By default,
it will install 8.11.0 on Ubuntu noble. The tzdata package prompts for input, so
use DEBIAN_FRONTEND=noninteractive when building the image.