Fix rpm upgrades to puppetcore
An earlier version of PR #757 assumed the default yum_source was changing to
yum-puppetcore.puppet.com. Later we decided not to introduce a breaking change
and instead use puppetcore* collections to opt-into the new behavior. However,
the PR wasn't updated to account for that. And I didn't notice the problem
because the docker upgrade script was using the puppet8 collection.

Change the upgrade script to use puppetcore* collections and update the redhat
and suse classes to install from yum-puppetcore.puppet.com when using the core
repos.

    ❯ export PUPPET_FORGE_TOKEN=...
    ❯ docker/bin/upgrade.sh rocky 7.34.0 8.11.0
    ...
    Notice: /Stage[main]/Puppet_agent::Install/Package[puppet-agent]/ensure: ensure changed '7.34.0-1.el8' to '8.11.0'
    ❯ docker/bin/upgrade.sh sles 7.34.0 8.11.0
    ...
    Notice: /Stage[main]/Puppet_agent::Install::Suse/Package[puppet-agent]/ensure: ensure changed '7.34.0-1.sles15' to '8.11.0'