(ITHELP-116185) Add fix dns step to Task Acceptance Tests workflow
In the Task Acceptance Tests workflow, Twingate was having trouble connecting
to Artifactory due to running the job in Docker containers. The Docker
containers used in the workflow run on Azure VM's and Microsoft adds their own
nameservers, 168.63.129.16. It gets copied to resolv.conf and interferes with
Twingate which caused DNS resolution failures with Artifactory.

To resolve this, this commit adds a "Fix DNS" step that removes the Azure DNS
name server and adds Twingate's DNS server.