(CAT-2413) Defaults to 8.14.0 for puppetcore agent version if non given
If puppetcore is selected and no version is given, defaults to version 8.14.0 for the Windows agent download.
As it is currently, if no version is given this will break as one is required by the download site in order to function.
This is intended as a temporary stop gap measure until such time as a permanent solution is found and implemented for the puppetcore downloads as a whole.