(maint) Update to PDK template 3.5.1
This commit updates this module to the latest PDK template (3.5.1).

The new template updates the TargetRubyVersion in .rubocop.yml to 3.1,
so this commit fixes all instances of Style/HashSyntax cop violations.

It also updates legacy facts from puppet_agent_spec, as legacy facts
were removed in newer versions of FacterDB.