(maint) Fail if puppetserver can't be started
By default, Beaker tests fail if they give a non-zero exit code.
However, when starting puppetserver in the master setup presuite, the
resource can fail and still return an exit code of 0.

This commit updates the start puppetserver step to use the --fail
option, which will return an exit code of 1 if the resource could not be
modified.